Disable hot_standby
Fix DB startup race condition causing error FATAL: the database system is starting up.

When the hot_standby parameter is set to true on a standby server, it will begin accepting connections once the recovery has brought the system to a consistent state. All such connections are strictly read-only; not even temporary tables may be written.

This is probably what was happening before: The server is coming up in hot_standby mode and depending on timing manage.py migrate tries to modify the data causing the error message.