Move Dockerfile and .dockerignore to top level
Frontend container is only necessary for development and once during build of the
production container. As we no longer have bind mounts in the production container
to the code outside we don't need to be able to recreate javascript bundle files
at runtime. These should be generated once when building the image and committed
to the docker image.

Moving the Dockerfile to the top-level allows for multi-stage
builds of the backend container enabling us to use a throwaway frontend container
to build a bundle and include this in the final backend container for a streamlined
production setup.

The separate frontend dockerfile is still included as this will
allow us to use it for local development and also possibly with a further override
file for bind mounting in a mixed development in production environment setup.