Switch to multi-stage-build for frontend
Removing the necessity for a frontend container for webpack.
Instead uses a multi-stage build and creates an intermediary
container in which the javascript bundles are built.

The created bundles are copied to the backend docker
during the build process and indexed. For production
frontend containers are no longer necessary and there
are no start-up-order issues anymore.

For development a frontend container is still used, as
webpack will watch and rebuild the bundle in development
mode as soon as a file is changed.

Co-authored-by: Kevin Yuan <kevin.yuan@gmx.de>