Rename .mjs files to .js
Unfortunately my IDE has some problems with the .mjs extensions and there is
currently no better work-around than renaming them to .js.

This shouldn't be a problem as .mjs is not strictly necessary. It only
serves to distinguish between JS modules and regular js files.
In our project we only use modules anyway.