Add mail handling examples (template editor, mail sending).
The main mail handling is done by django-post-office. Mail templates are defined by the contacting group (e.g. hospitals), details such as name or personalized unsubscribe link are added / rendered on delivery. All mails are logged in case of unexpected issues with mail delivery (say a deleted sendgrid api key). Mail templates are saved to the database, a user could load an old template to reuse a written message.
A mail preview is shown before sending the mail to show the user what the actual send out mail would look like.
Mails are send in form of multipart/alternative mails, a text version of the mail is created (without html tags but with linebreaks etc).
Mail delivery is handled asynchronous, per default they are queued and sending mails has to be either manually specified (executing a command) or can be handled by celery (will be added later on).
This version only contains an editor example where the user add a new template, preview it and send mails to a hardcoded number of users. This example needs to be extended to send mails to an "matching group".
The mail app also probably needs to be merged mack into the mail matching app.